Implemented fseek(), fstat(), and fclose() in Fat16 and VFS

+++ src/disk/diskreader.rs
@@ -5,8 +5,8 @@ use crate::status::ErrorCode;
 use super::ata_pio::AtaPio;
 
 pub trait DiskReader: Send + Sync {
-    fn read(&self, lba: usize, out: &mut [u16], total: usize) -> Result<usize, ErrorCode>;
-    fn write(&self, lba: usize, data: &mut [u16]) -> Result<(), ErrorCode>;
+    fn read(&self, lba: usize, out: &mut [u8], total: usize) -> Result<usize, ErrorCode>;
+    fn write(&self, lba: usize, data: &mut [u8]) -> Result<(), ErrorCode>;
     fn resolve(index: usize) -> Result<Self, ErrorCode>
     where
         Self: Sized;
